TP-32 FIX

+++ TradingCms/TradingCms.Data/OrdersToProducts.cs
@@ -5,6 +5,6 @@ namespace TradingCms.Data
     {
         public virtual Order Order { get; set; }
         public virtual Product Product { get; set; }
-        public virtual int Price { get; set; }
+        public virtual decimal Price { get; set; }
     }
 }
